Rename client/src/pages to client/src/layouts
Update imports as needed
