+++ client/src/stories/user-auth.pages.stories.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { storiesOf } from '@storybook/react';
 import { StaticRouter as Router } from 'react-router-dom';
 
-import Login from '../pages/Login/Login';
+import Login from '../layouts/Login/Login';
 
 storiesOf('User Auth Pages | Login', module)
   .add('it renders', () => (
